pyproject.toml
[tool.poetry]
name = "financez"
version = "0.1.0"
description = ""
authors = ["serge"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.3"
djangorestframework = "^3.15.1"
gunicorn = "^20.1.0"
poetryup = "^0.12.7"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.13"
django-cors-headers = "^3.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
